Add README and update main.c and main.h for wait time calculation improvements

+++ main.h
@@ -12,7 +12,7 @@ extern pthread_t threads_ids[NUM_THREADS];
 extern pthread_mutex_t mutex;
 extern sem_t parking_semaphore;
 extern int total_cars;
-extern int total_wait_time;
+extern double total_wait_time;
 
 void *access_parking_lot(void *arg);
 void log_event(const char *event, int thread_id, time_t event_time);
